services/publisher_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
httpx>=0.25.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
asyncpg>=0.29.0
python-dotenv>=1.0.0
boto3>=1.34.0
botocore>=1.34.0
openai>=1.3.0  # For GPT-4 API
python-wordpress-xmlrpc>=2.3  # For WordPress API
aiofiles>=23.2.1
google-generativeai>=0.3.0  # For Google Gemini API